Extracted duplicates to abstract base class
