+++ crawler/src/main/java/com/javaacademy/crawler/jsoup/CzytamScrapper.java
@@ -34,17 +34,10 @@ public class CzytamScrapper extends JsoupBookScrapper {
     }
 
     @Override
-    Set<BookModel> parseSingleGrid() {
+    Set<String> getLinksFromGrid() {
         Elements elements = getDoc().select("h3.product-title > a");
         Set<String> sublinks = new HashSet<>(elements.eachAttr("href"));
-        Set<String> links = new HashSet<>(sublinks.stream().map(BASE_URL::concat).collect(Collectors.toSet()));
-        Set<BookModel> bookModels = new HashSet<>();
-        for (String link : links) {
-            connect(link);
-            BookModel bookModel = parseSinglePage(link);
-            bookModels.add(bookModel);
-        }
-        return bookModels;
+        return new HashSet<>(sublinks.stream().map(BASE_URL::concat).collect(Collectors.toSet()));
     }
 
     @Override
@@ -93,10 +86,5 @@ public class CzytamScrapper extends JsoupBookScrapper {
         Elements elements = getDoc().select("div.price > strong");
         return elements.isEmpty() ? 0 :
                 parsePrice(elements.first().text().replace(",", ".").replace(" PLN", ""));
-    }
-
-    @Override
-    String getLink(Element element) {
-        return null;
     }
 }
