+++ crawler/src/main/java/com/javaacademy/crawler/jsoup/MatrasScrapper.java
@@ -37,17 +37,11 @@ public class MatrasScrapper extends JsoupBookScrapper {
     }
 
     @Override
-    Set<BookModel> parseSingleGrid() {
+    Set<String> getLinksFromGrid() {
         Elements elements = getDoc().select("div.row.row-items span.right-side a");
         Set<String> links = new HashSet<>(elements.eachAttr("href"));
         links.removeIf(s -> s.startsWith(BASE_URL + "/szukaj/"));
-        Set<BookModel> bookModels = new HashSet<>();
-        for (String link : links) {
-            connect(link);
-            BookModel bookModel = parseSinglePage(link);
-            bookModels.add(bookModel);
-        }
-        return bookModels;
+        return links;
     }
 
     @Override
@@ -98,10 +92,5 @@ public class MatrasScrapper extends JsoupBookScrapper {
     double getRetailPrice() {
         Elements elements = getDoc().select("div.this-main-price");
         return elements.isEmpty() ? 0 : parsePrice(elements.attr("content"));
-    }
-
-    @Override
-    String getLink(Element element) {
-        return element.select(".cover").select("a").attr("href");
     }
 }
